Renamed game/items to generic stones n gems
